+++ WallApp/UI/Models/SettingsModel.cs
@@ -3,7 +3,7 @@
     public abstract class SettingsModel
     {
         public virtual int FrameRate { get; set; }
-        public virtual double BackBufferScale { get; set; }
+        public virtual float BackBufferScale { get; set; }
         public abstract int GetModuleCount();
         public abstract int GetLayerCount();
